API Fixes

+++ ersilia/hub/fetch/register/standard_example.py
@@ -110,11 +110,17 @@ class ModelStandardExample(ErsiliaBase):
             self.logger.debug("The environment name is {0}".format(env_name))
             SimpleConda().run_commandlines(env_name, commands)
 
-        self.logger.info(f"Run log: {open(run_log).read()}")
+        # Check if log file exists before trying to read it
+        if os.path.exists(run_log):
+            self.logger.info(f"Run log: {open(run_log).read()}")
+        else:
+            self.logger.warning(f"Run log file {run_log} was not created")
+
         self._check_file_exists(output_csv=output_csv)
         is_fetched_successfully = self._validate_csv(output_csv)
         self.logger.debug("Removing log file: {0}".format(run_log))
-        os.remove(run_log)
+        if os.path.exists(run_log):
+            os.remove(run_log)
         if not is_fetched_successfully:
             echo("The model is getting deleted due to it produced all empty values!")
             run_command(f"ersilia -v delete {self.model_id}")
